requirements.txt
streamlit
sentence-transformers
faiss-cpu
gpt4all
pymupdf